Remove non-compliant instances
The instances for lists, sequences and vectors do not obey the specified laws. The reason is that indices for these types are not persistent: they are assigned sequentially starting from 0, and that means that filtering can change the indexing.

The law is given as

    ifilter f . ifilter g ≡ ifilter (i -> liftA2 (&&) (f i) (g i))

but that doesn't hold for any of these. Take for example

    f i _ = even i
    g i _ = even i

and evaluate both sides on `[0..10]`. Then the left side will be `[0,4,8]` (because it removes the odd indices, and then removes the odd indices again) but the right side will be `[0,2,4,6,8,10]`.

The law is a sensible one, and there is a serious danger that someone will use rewrite rules which assume it, which would have disastrous results for these types.