Removing long-deprecated Option
The name `Option` is a very natural one, and consequently has been used to refer to several different types. At the moment, a reader might assume that it means a strict version of `Maybe` (from the `strict-data` package). However, these instances actually refer to a version of `Maybe` that was in `Data.Semigroup`.

According to the documentation for base-4.14.2.0, "In GHC 8.4 and higher, the Monoid instance for Maybe has been corrected to lift a Semigroup instance instead of a Monoid instance. Consequently, this type is no longer useful. It will be marked deprecated in GHC 8.8 and removed in GHC 8.10."

We should no longer provide instances for a long-removed type.